:feat adoption redo :fix proband select
re-worked how adoption and adoption types [adopted_in|adopted_out] are stored. This is to better represent logic of allowing adopted-out parents to be added, if known [feature setup but not ready].

Also added in logic to allow the proband to be de-selected another person to be added as proband.

:bug multiple individuals can be proband now.  Need to add in logic to remove proband tag for other individuals.  Also need to move proband/consultand checkboxes to a section.  Logic to only allow 1 of proband|consutand. And note to user if no proband is in the  pedigree